refactor: 폴더 구조 변경, final 붙이기
