+++ src/main/java/kr/pickple/back/member/service/MemberService.java
@@ -11,8 +11,8 @@ import kr.pickple.back.address.dto.response.MainAddressResponse;
 import kr.pickple.back.address.service.AddressService;
 import kr.pickple.back.member.domain.Member;
 import kr.pickple.back.member.domain.MemberPosition;
-import kr.pickple.back.member.dto.AuthenticatedMemberResponse;
-import kr.pickple.back.member.dto.MemberCreateRequest;
+import kr.pickple.back.member.dto.request.MemberCreateRequest;
+import kr.pickple.back.member.dto.response.AuthenticatedMemberResponse;
 import kr.pickple.back.member.exception.MemberException;
 import kr.pickple.back.member.repository.MemberPositionRepository;
 import kr.pickple.back.member.repository.MemberRepository;
